Column-width and transition adjustments
Adjusted the width of most of the text columns to account for really
wide screens in order to keep them within the 7-10 words-per-line
recommendation for readability. Also added a new 'transit' plugin that
takes the place of jquery's '.animate' function for smoother
transitions. Also, the top timelapse and the flyover now have their
height determined, not by measuring their heights after the document
loads, but by finding the browser width and doing this "var videoheight
= Math.floor(videowidth * .42);"